requirements.txt
appnope==0.1.0
bleach==2.0.0
decorator==4.0.11
Django==1.10.4
djangorestframework==3.5.3
djangotoolbox==1.8.0
entrypoints==0.2.2
feedparser==5.2.1
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
mongoengine==0.8.7
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==2.8
python-dateutil==2.6.0
pyzmq==16.0.2
qtconsole==4.2.1
simplegeneric==0.8.1
singlemodeladmin==0.6
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
